Use histograms and boxplots to get an general idea of what data looks like

Creates a subplot for each specified feature. Each subplot can be histogram or a boxplot. Same as the generate.descriptive.plots() function except this function saves the plot to a pdf file.

generate.descriptive.plots.save.pdf(
  inputted.data,
  overall.plot.layout,
  plot.type,
  plot.features,
  plot.name,
  width = 5,
  height = 8
)

Arguments

inputted.data
A dataframe.

overall.plot.layout
A vector with two elements to specify how many rows and columns to include in the par plot. c(rows, cols)

plot.type
A vector that specifies, for each subplot, should features be displayed in histogram or boxplot form. Elements can be "histogram" or "boxplot".

plot.features
A vector that specifies, for each subplot, what features should be analyzed. The elements should be names of columns that contain continous numerical data.

plot.name
A string for the name of the plot. Will be used for the name of the output pdf.

width
Number indicating the width of plot. Default is 5.

height
Number indicating the width of the plot. Default is 8.

Value

No object is outputted, but a plot is saved as a pdf.

See also